Hey Mr. Kennedy you big bitch, this project sucked
Worse implementation that doesn't use UniqueMeregdIterator class that I made. It also doesn't implement the 'required features' for the extra credit. Every value is still actually in the buffer, nothing actually got deleted. I got a 0/10 on tests again because I couldn't submit for 3 days. I wasted tons of my submissions on your stupid typos, and flat our errors in syllabus, and now I have to delete a shit ton of code and work because an absolute misrepresentation in how you liked the extra credit